+++ R/attributes.R
@@ -473,26 +473,34 @@ set_all_borders <- function(ht, row, col, value, byrow = FALSE) {
 #' bottom, left and right of a group of cells.
 #'
 #' @param ht A huxtable
-#' @param row A set of rows. See below.
-#' @param col A set of columns.
+#' @param row A row specifier. See \code{\link{rowspecs}} for details.
+#' @param col A column specifier.
 #' @param value A numeric value for the border width. Set to 0 for no border.
 #'
 #' @return The modified huxtable.
-#' @details
-#' Only standard R subsetting may be used for \code{row} and \code{col}. That is,
-#' logical, numeric or character indices are allowed, but not the tricks in \code{\link{rowspecs}}.
-#'
 #' @export
 #'
 #' @seealso \code{\link{left_border}}, \code{\link{set_all_borders}}
 #' @examples
 #' ht <- huxtable(a = 1:3, b = 1:3)
-#' ht <- set_outer_borders(ht, 2:3, 1:2, 1)
-#' ht
+#' set_outer_borders(ht, 1)
+#' set_outer_borders(ht, 2:3, 1:2, 1)
+#'
 set_outer_borders <- function(ht, row, col, value) {
-  if (is.character(row)) row <- rownames(ht) %in% row
+  if (missing(col) && missing(value)) {
+    value <- row
+    row <- seq_len(nrow(ht))
+    col <- seq_len(ncol(ht))
+  } else if (missing(value)) {
+    value <- col
+    if (!is.matrix(row)) stop("No columns specified, but `row` argument did not evaluate to a matrix")
+    # row is a 2-matrix of row, col vectors;
+    col <- seq(min(row[, 2]), max(row[, 2]))
+    row <- seq(min(row[, 1]), max(row[, 1]))
+  }
+  row <- get_rc_spec(ht, row, 1)
+  col <- get_rc_spec(ht, col, 2)
   if (is.logical(row)) row <- which(row)
-  if (is.character(col)) col <- colnames(ht) %in% col
   if (is.logical(col)) col <- which(col)
 
   left_border(ht)[row, min(col)]    <- value